重构代码结构

+++ deploy_code.py
@@ -4,11 +4,12 @@ import shutil
 import os
 
 # python -m py_compile file.py
-src_dir = 'pretrain_scripts'
+src_dir = 'src'
 dist_dir = 'dist'
 
 # clean
-shutil.rmtree(dist_dir)
+if os.path.exists(dist_dir):
+    shutil.rmtree(dist_dir)
 os.mkdir(dist_dir)
 
 files = os.listdir(src_dir)
@@ -35,5 +36,4 @@ for file in files:
     print('dist: %s -> %s' % (oldfile, newfile))
     shutil.copy(oldfile, newfile)
 
-shutil.copy('config.py', dist_dir)
-shutil.copy('config.json', dist_dir)
+shutil.copy('config/config.json', dist_dir)
